+++ drivers/mmc/hi6220_dw_mmc.c
@@ -164,13 +164,13 @@ static const struct hisi_mmc_data hi3798mv2x_mmc_data = {
 
 static const struct udevice_id hi6220_dwmmc_ids[] = {
 	{ .compatible = "hisilicon,hi6220-dw-mshc",
-	  .data = (ulong)&hi6220_mmc_data },
+	  .data = (uintptr_t)&hi6220_mmc_data },
 	{ .compatible = "hisilicon,hi3798cv200-dw-mshc",
-	  .data = (ulong)&hi6220_mmc_data },
+	  .data = (uintptr_t)&hi6220_mmc_data },
 	{ .compatible = "hisilicon,hi3798mv200-dw-mshc",
-	  .data = (ulong)&hi3798mv2x_mmc_data },
+	  .data = (uintptr_t)&hi3798mv2x_mmc_data },
 	{ .compatible = "hisilicon,hi3660-dw-mshc",
-	  .data = (ulong)&hi3660_mmc_data },
+	  .data = (uintptr_t)&hi3660_mmc_data },
 	{ }
 };
