+++ drivers/net/sun8i_emac.c
@@ -907,17 +907,17 @@ static const struct emac_variant emac_variant_h6 = {
 
 static const struct udevice_id sun8i_emac_eth_ids[] = {
 	{ .compatible = "allwinner,sun8i-a83t-emac",
-	  .data = (ulong)&emac_variant_a83t },
+	  .data = (uintptr_t)&emac_variant_a83t },
 	{ .compatible = "allwinner,sun8i-h3-emac",
-	  .data = (ulong)&emac_variant_h3 },
+	  .data = (uintptr_t)&emac_variant_h3 },
 	{ .compatible = "allwinner,sun8i-r40-gmac",
-	  .data = (ulong)&emac_variant_r40 },
+	  .data = (uintptr_t)&emac_variant_r40 },
 	{ .compatible = "allwinner,sun8i-v3s-emac",
-	  .data = (ulong)&emac_variant_v3s },
+	  .data = (uintptr_t)&emac_variant_v3s },
 	{ .compatible = "allwinner,sun50i-a64-emac",
-	  .data = (ulong)&emac_variant_a64 },
+	  .data = (uintptr_t)&emac_variant_a64 },
 	{ .compatible = "allwinner,sun50i-h6-emac",
-	  .data = (ulong)&emac_variant_h6 },
+	  .data = (uintptr_t)&emac_variant_h6 },
 	{ }
 };
