+++ drivers/pci/pcie_dw_imx.c
@@ -335,8 +335,8 @@ static const struct dm_pci_ops pcie_dw_imx_ops = {
 };
 
 static const struct udevice_id pcie_dw_imx_ids[] = {
-	{ .compatible = "fsl,imx8mm-pcie", .data = (ulong)&imx8mm_chip_info, },
-	{ .compatible = "fsl,imx8mp-pcie", .data = (ulong)&imx8mp_chip_info, },
+	{ .compatible = "fsl,imx8mm-pcie", .data = (uintptr_t)&imx8mm_chip_info, },
+	{ .compatible = "fsl,imx8mp-pcie", .data = (uintptr_t)&imx8mp_chip_info, },
 	{ }
 };
